refactor(selfhost): move Package/Resolve/CfgEnv types to selfhost/api
Follow-up to PR #761. Moves the remaining cross-boundary data types
into internal/selfhost/api so the whole public-type surface now lives
in one subpackage. Consumers (cmd/osty, internal/check, native-checker
binary) can depend on api alone for types; selfhost re-exports each
type via a Go type alias so existing callers compile unchanged.

Moved:
- Package*: PackageCheckFile, PackageCheckGenericBound, PackageCheckFn,
  PackageCheckField, PackageCheckVariant, PackageCheckAlias,
  PackageCheckType, PackageCheckInterfaceExt, PackageCheckImport,
  PackageCheckInput
- Resolve: PackageResolveFile (alias), PackageResolveInput, UseEdge,
  PackageUses, WorkspaceUses, CycleDiag
- CfgEnv (was: `selfhost.CfgEnv`) — the `toSelf()` method was replaced
  with a free function `cfgEnvToSelf(c *CfgEnv) *SelfResolveCfgEnv` so
  the type itself has no selfhost-internal references.

No call-site churn; all existing `selfhost.X` references resolve through
the alias.
